requirements.txt
fastapi~=0.76.0
uvicorn[standard]~=0.17.6
PyJWT~=2.3.0
mysql-connector-python~=8.0.29
pydantic
python-dateutil